requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/dev.in
#
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via
    #   -r requirements/prod.txt
    #   django
attrs==21.2.0
    # via pytest
backcall==0.2.0
    # via ipython
backports.entry-points-selectable==1.1.0
    # via virtualenv
black==21.7b0
    # via -r requirements/dev.in
certifi==2021.5.30
    # via
    #   -r requirements/prod.txt
    #   requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.4
    # via
    #   -r requirements/prod.txt
    #   requests
click==8.0.3
    # via
    #   black
    #   pip-tools
decorator==5.1.0
    # via ipython
distlib==0.3.3
    # via virtualenv
django==3.2.6
    # via
    #   -r requirements/prod.txt
    #   django-admin-numeric-filter
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   model-bakery
django-admin-numeric-filter==0.1.6
    # via -r requirements/prod.txt
django-cors-headers==3.8.0
    # via -r requirements/prod.txt
django-extensions==3.1.3
    # via -r requirements/prod.txt
django-filter==2.4.0
    # via -r requirements/prod.txt
djangorestframework==3.12.4
    # via
    #   -r requirements/prod.txt
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.7.2
    # via -r requirements/prod.txt
drf-writable-nested==0.6.3
    # via -r requirements/prod.txt
filelock==3.3.1
    # via virtualenv
flake8==3.9.2
    # via -r requirements/dev.in
gunicorn==20.1.0
    # via -r requirements/prod.txt
identify==2.3.1
    # via pre-commit
idna==3.2
    # via
    #   -r requirements/prod.txt
    #   requests
iniconfig==1.1.1
    # via pytest
ipython==7.27.0
    # via -r requirements/dev.in
isort==5.9.3
    # via -r requirements/dev.in
jedi==0.18.0
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via flake8
model-bakery==1.3.2
    # via -r requirements/dev.in
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
packaging==21.0
    # via pytest
parso==0.8.2
    # via jedi
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.1
    # via -r requirements/prod.txt
pip-tools==6.2.0
    # via -r requirements/dev.in
platformdirs==2.4.0
    # via virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.14.0
    # via -r requirements/dev.in
prompt-toolkit==3.0.21
    # via ipython
psycopg2-binary==2.9.1
    # via -r requirements/prod.txt
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pygments==2.10.0
    # via ipython
pyjwt==2.1.0
    # via
    #   -r requirements/prod.txt
    #   djangorestframework-simplejwt
pyparsing==3.0.2
    # via packaging
pytest==6.2.5
    # via
    #   pytest-django
    #   pytest-mock
pytest-django==4.4.0
    # via -r requirements/dev.in
pytest-mock==3.6.1
    # via -r requirements/dev.in
python-decouple==3.4
    # via -r requirements/prod.txt
pytz==2021.1
    # via
    #   -r requirements/prod.txt
    #   django
pyyaml==6.0
    # via pre-commit
regex==2021.10.23
    # via black
requests==2.26.0
    # via -r requirements/prod.txt
six==1.16.0
    # via virtualenv
sqlparse==0.4.1
    # via
    #   -r requirements/prod.txt
    #   django
toml==0.10.2
    # via
    #   pre-commit
    #   pytest
tomli==1.2.2
    # via
    #   black
    #   pep517
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.6
    # via
    #   -r requirements/prod.txt
    #   requests
virtualenv==20.9.0
    # via pre-commit
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
